big update fuck les types

+++ src/models/userModel.ts
@@ -69,7 +69,7 @@ export async function getUserPrivateData(username: string): Promise<UserPrivateD
     }
 }
 
-export async function getUserCreate(username: String): Promise<UserCreate | null> {
+export async function getUserCreate(username: string): Promise<UserCreate | null> {
     const user: UserCreate | null = await prisma.user.findUnique({
         where: { username: username }
     });
@@ -99,13 +99,15 @@ export async function createUser(newUser: UserCreate): Promise<UserPrivateData |
 }
 
 export async function deleteUser(username: string): Promise<Boolean> {
-    const userId = getPlayerIdByUsername(username);
-    await prisma.game.deleteMany({
-        where: { playerId: userId }
-    })
-    await prisma.user.delete({
-        where: { username: username }
-    })
+    const userId = await getPlayerIdByUsername(username);
+    if (userId) {
+        await prisma.game.deleteMany({
+            where: { playerId: userId }
+        })
+        await prisma.user.delete({
+            where: { username: username }
+        })
+    }
     return true;
 }
 
@@ -121,7 +123,7 @@ export async function updatePlayedGame(username: string): Promise<UserPublicData
     return user;
 }
 
-export async function getPlayerIdByUsername(username: string): Promise<number> {
+export async function getPlayerIdByUsername(username: string): Promise<number | null> {
     const user_id = await prisma.user.findUnique({
         where: {
             username: username
@@ -130,7 +132,10 @@ export async function getPlayerIdByUsername(username: string): Promise<number> {
             id: true
         }
     });
-    return user_id.id;
+    if (user_id)
+        return user_id.id;
+    else
+        return null
 }
 
 export async function getUserAllData(id: number) {
@@ -149,7 +154,7 @@ export async function disableUser(id: number): Promise<Boolean> {
         where: { id: id },
         select: { disabled: true }
     });
-
+    //@ts-ignore
     const newDisabledState = !user.disabled;
 
     await prisma.user.update({
@@ -165,5 +170,6 @@ export async function getAdminField(id: number): Promise<Boolean> {
         where: { id: id },
         select: { admin: true }
     });
+    //@ts-ignore
     return userAdmin.admin;
 }
